+++ src/onegov.election_day/onegov/election_day/forms/election.py
@@ -3,7 +3,7 @@ from onegov.election_day import _
 from onegov.form import Form
 from wtforms import IntegerField, RadioField, StringField
 from wtforms.fields.html5 import DateField, URLField
-from wtforms.validators import NumberRange, InputRequired
+from wtforms.validators import NumberRange, InputRequired, Optional
 
 
 class ElectionForm(Form):
@@ -66,43 +66,13 @@ class ElectionForm(Form):
         label=_("Related link")
     )
 
-    upload_type = RadioField(
-        _("Upload"),
-        choices=[
-            ('manual', _("Manual")),
-            ('wabsti', _("Automatic (Wabsti)")),
-        ],
+    absolute_majority = IntegerField(
+        label=_("Absolute majority"),
         validators=[
-            InputRequired()
+            Optional(),
+            NumberRange(min=1)
         ],
-        default='manual'
-    )
-
-    upload_wabsti_district = StringField(
-        label=_("Automatic upload (Wabsti): 'SortWahlkreis'"),
-        validators=[
-            InputRequired()
-        ],
-        render_kw=dict(force_simple=True),
-        depends_on=('upload_type', 'wabsti'),
-    )
-
-    upload_wabsti_number = StringField(
-        label=_("Automatic upload (Wabsti): 'SortGeschaeft'"),
-        validators=[
-            InputRequired()
-        ],
-        render_kw=dict(force_simple=True),
-        depends_on=('upload_type', 'wabsti'),
-    )
-
-    upload_wabsti_folder = StringField(
-        label=_("Automatic upload (Wabsti): Folder"),
-        validators=[
-            InputRequired()
-        ],
-        render_kw=dict(force_simple=True),
-        depends_on=('upload_type', 'wabsti')
+        depends_on=('election_type', 'majorz'),
     )
 
     def set_domain(self, principal):
@@ -117,6 +87,7 @@ class ElectionForm(Form):
         model.type = self.election_type.data
         model.shortcode = self.shortcode.data
         model.number_of_mandates = self.mandates.data
+        model.absolute_majority = self.absolute_majority.data
 
         model.title_translations = {}
         model.title_translations['de_CH'] = self.election_de.data
@@ -149,6 +120,7 @@ class ElectionForm(Form):
         self.shortcode.data = model.shortcode
         self.election_type.data = model.type
         self.mandates.data = model.number_of_mandates
+        self.absolute_majority.data = model.absolute_majority
 
         meta_data = model.meta or {}
         self.related_link.data = meta_data.get('related_link', '')